lib: rename tpm_kdfa to tpm2_kdfa in order to be consistent
