+++ lib/tpm2_kdfa.c
@@ -31,7 +31,7 @@
 #include <openssl/hmac.h>
 
 #include "log.h"
-#include "tpm_kdfa.h"
+#include "tpm2_kdfa.h"
 #include "tpm2_openssl.h"
 #include "tpm2_util.h"
