+++ lib/tpm2_kdfa.h
